refactor: simplify repo structure
feat: add screening abstracts repo
