+++ src/sr_assistant/core/schemas.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+import typing as t
 
 from pydantic import BaseModel, Field
 
-if TYPE_CHECKING:
+if t.TYPE_CHECKING:
     from pydantic.types import PositiveInt
+
+    from sr_assistant.core.types import ExclusionReasonType
 
 
 class EventDetailBase(BaseModel):
@@ -61,4 +63,34 @@ class EventDetailBase(BaseModel):
         default=1,
         description="Event version. Must be updated in corresponding event detail schema if schema mutated.",
         title="Event Version",
+    )
+
+# The JSONSchema of this model is passed as a tool-call schema for the model.
+# OpenAI has specific limitations, like we can't set a minimum or maximum value for the
+# confidence_score and have to rely on the description.
+class ScreeningResponse(BaseModel):
+    """Your systematic review screening response/decision for the given study.
+
+    Consider the given context from the protocol carefully. Screening abstracts can be
+    tricky, so assign confidence score and decision accordingly.
+    """
+    decision: t.Literal["include", "exclude", "uncertain"] = Field(
+        ...,
+        description="The systematic review abstracts screening decision. Should this study be included or not? Or are you uncertain? If your confidence score is below 0.8, assign 'uncertain'.",
+    )
+    confidence_score: float = Field(
+        ...,
+        description="The confidence score for the decision. [0.0, 1.0].",
+    )
+    rationale: str = Field(
+        ...,
+        description="The rationale for the decision. Be specific. Explainable AI is imporant. Don't just reiterate the category/categories, but explain how and why you reached this conclusion.",
+    )
+    extracted_quotes: list[str] | None = Field(
+        default=None,
+        description="Supporting quotes from the title/abstract. Can be omitted if uncertain.",
+    )
+    exclusion_reason_categories: list[ExclusionReasonType] | None = Field(
+        default=None,
+        description="Omit if the decision is 'include'. If the decision is 'exclude' or 'uncertain', The PRISMA exclusion reason categories for the decision. This complements the 'rationale' field.",
     )
