--- src/sr_assistant/core/schemas/events.py
@@ -1,1 +1,0 @@
-from __future__ import annotations
